requirements.txt
python-dotenv==1.0.0
stripe==7.4.0
flask==2.3.3
openai==1.3.0
python-telegram-bot==20.6 